requirements.txt
tensorflow==2.13.0
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
werkzeug==2.3.7
